feat: imporoved drag and drop component
