+++ src/components/index.ts
@@ -17,3 +17,4 @@ export * from './HomeHeroCard';
 export * from './HomeListings';
 export * from './PaymentForm';
 export * from './StripeContainer';
+export * from './DragAndDrop';
