+++ src/sections/Host/Host.tsx
@@ -1,39 +1,12 @@
 import { Button, Form, Input, InputNumber, Radio } from 'antd';
 import { Content } from 'antd/es/layout/layout';
+import { useState } from 'react';
 import { BsHouseDoorFill } from 'react-icons/bs';
 import { MdApartment } from 'react-icons/md';
-import { BsPlusLg } from 'react-icons/bs';
-import React, { useState } from 'react';
+import { DragAndDrop } from '../../components';
 
 export const Host = () => {
   const [image, setImage] = useState<string | ArrayBuffer | null>(null);
-  const [drag, setDrag] = useState<boolean>(false);
-
-  const toBase64 = (file: File): Promise<string | ArrayBuffer | null> =>
-    new Promise((resolve, reject) => {
-      const reader = new FileReader();
-      reader.readAsDataURL(file);
-      reader.onload = () => resolve(reader.result);
-      reader.onerror = (error) => reject(error);
-    });
-
-  const dragStartHandler = (e: any) => {
-    e.preventDefault();
-    setDrag(true);
-  };
-
-  const dragLeaveHandler = (e: any) => {
-    e.preventDefault();
-    setDrag(false);
-  };
-
-  const dropHandler = async (e: any) => {
-    e.preventDefault();
-    const files = [...e.dataTransfer.files];
-    const fileBase64 = await toBase64(files[0]);
-    setImage(fileBase64);
-    setDrag(false);
-  };
 
   const onFinish = async (values: any) => {
     values.image = image;
@@ -99,39 +72,11 @@ export const Host = () => {
           rules={[{ required: true, message: 'This field can not be empty!' }]}>
           <InputNumber />
         </Form.Item>
-        <div className='flex items-center'>
-          <div className='flex flex-col items-center w-28'>
-            <h4 className='text-center mb-2'>Drop your avatar here</h4>
-            <div
-              className='
-              w-28
-              h-28
-              bg-gray-100
-              border-xl
-              border-2
-              border-gray-200
-              rounded
-              flex
-              items-center
-              justify-center'
-              onDragStart={dragStartHandler}
-              onDragLeave={dragLeaveHandler}
-              onDragOver={dragStartHandler}
-              onDrop={dropHandler}>
-              {drag ? (
-                <span className='animate-bounce'>Drop here</span>
-              ) : (
-                <BsPlusLg size='2em' />
-              )}
-            </div>
-          </div>
-          {image ? (
-            <img
-              src={image.toString()}
-              className='h-40 ml-10 rounded'
-            />
-          ) : null}
-        </div>
+        <DragAndDrop
+          image={image}
+          setImage={setImage}
+          maxFileSize={500000}
+        />
         <Button
           htmlType='submit'
           type='primary'
